+++ trixnity-client/src/commonTest/kotlin/net/folivo/trixnity/client/key/DehydratedDeviceServiceTest.kt
@@ -11,6 +11,7 @@ import kotlinx.datetime.Clock
 import kotlinx.serialization.ExperimentalSerializationApi
 import net.folivo.trixnity.client.*
 import net.folivo.trixnity.client.crypto.ClientOlmStore
+import net.folivo.trixnity.client.mocks.KeyServiceMock
 import net.folivo.trixnity.client.mocks.SignServiceMock
 import net.folivo.trixnity.client.store.KeySignatureTrustLevel
 import net.folivo.trixnity.client.store.KeySignatureTrustLevel.Valid
@@ -62,6 +63,7 @@ class DehydratedDeviceServiceTest : TrixnityBaseTest() {
     private val aliceDevice = "ALICEDEVICE"
 
     private val signServiceMock = SignServiceMock()
+    private val keyServiceMock = KeyServiceMock()
     private val keyStore = getInMemoryKeyStore()
     private val olmCryptoStore = getInMemoryOlmStore()
     private val olmStore = ClientOlmStore(
@@ -97,8 +99,9 @@ class DehydratedDeviceServiceTest : TrixnityBaseTest() {
         json = json,
         olmStore = olmStore,
         signService = signServiceMock,
+        keyService = keyServiceMock,
         clock = clock,
-        matrixClientServerApi = matrixClientConfiguration,
+        config = matrixClientConfiguration,
     )
 
     @Test
@@ -132,7 +135,7 @@ class DehydratedDeviceServiceTest : TrixnityBaseTest() {
         }
 
         val dehydrateDeviceJob = launch {
-            cut.dehydrateDevice(dehydratedDeviceKey)
+            cut.tryDehydrateDevice(dehydratedDeviceKey)
         }
         delay(100.milliseconds)
         dehydrateDeviceJob.isActive shouldBe true
@@ -147,6 +150,21 @@ class DehydratedDeviceServiceTest : TrixnityBaseTest() {
         delay(100.milliseconds)
         dehydrateDeviceJob.isActive shouldBe true
         keyStore.updateOutdatedKeys { setOf() }
+        delay(100.milliseconds)
+        dehydrateDeviceJob.isActive shouldBe true
+        keyStore.updateCrossSigningKeys(alice) {
+            setOf(
+                StoredCrossSigningKeys(
+                    SignedCrossSigningKeys(
+                        CrossSigningKeys(
+                            alice, setOf(CrossSigningKeysUsage.SelfSigningKey), keysOf(
+                                Ed25519Key("wrong public key", "wrong public key")
+                            )
+                        ), mapOf()
+                    ), Valid(true)
+                )
+            )
+        }
         delay(100.milliseconds)
         dehydrateDeviceJob.isActive shouldBe true
         keyStore.updateCrossSigningKeys(alice) {
@@ -297,7 +315,7 @@ class DehydratedDeviceServiceTest : TrixnityBaseTest() {
 
         dehydratedDeviceAccount.markKeysAsPublished()
 
-        cut.rehydrateDevice(dehydratedDeviceKey)
+        cut.tryRehydrateDevice(dehydratedDeviceKey)
 
         getDehydratedDeviceCalled shouldBe true
         getDehydratedDeviceEventsCalled shouldBe true
@@ -325,8 +343,128 @@ class DehydratedDeviceServiceTest : TrixnityBaseTest() {
             }
         }
 
-        cut.rehydrateDevice("testKey".encodeToByteArray())
+        cut.tryRehydrateDevice("testKey".encodeToByteArray())
 
         getDehydratedDeviceCalled shouldBe true
+    }
+
+    @Test
+    fun `rehydrateDevice should ignore wrong key`() = runTest {
+        val dehydratedDeviceKey = SecureRandom.nextBytes(32)
+        val deviceId = "DEHYDRATED_DEVICE_ID"
+        var getDehydratedDeviceCalled = false
+        var getDehydratedDeviceEventsCalled = false
+
+        val dehydratedDeviceAccount = OlmAccount.create()
+        dehydratedDeviceAccount.generateOneTimeKeys(dehydratedDeviceAccount.maxNumberOfOneTimeKeys)
+
+        val bobAccount = OlmAccount.create()
+
+        val encryptedData = encryptAesHmacSha2(
+            content = dehydratedDeviceAccount.pickle("").encodeToByteArray(),
+            key = dehydratedDeviceKey,
+            name = DehydratedDeviceData.DehydrationV2Compatibility.ALGORITHM
+        )
+
+        val deviceData = DehydratedDeviceData.DehydrationV2Compatibility(
+            iv = encryptedData.iv,
+            encryptedDevicePickle = encryptedData.ciphertext,
+            mac = encryptedData.mac
+        )
+
+        val megolmSession = freeAfter(OlmOutboundGroupSession.create()) { outboundSession ->
+            RoomKeyEventContent(
+                roomId = roomId,
+                sessionId = outboundSession.sessionId,
+                sessionKey = outboundSession.sessionKey,
+                algorithm = EncryptionAlgorithm.Megolm,
+            )
+        }
+        val encryptedMessage = freeAfter(
+            OlmSession.createOutbound(
+                bobAccount,
+                dehydratedDeviceAccount.identityKeys.curve25519,
+                dehydratedDeviceAccount.oneTimeKeys.curve25519.values.first()
+            )
+        ) { olmSession ->
+            olmSession.encrypt(
+                json.encodeToString(
+                    decryptedOlmEventSerializer,
+                    DecryptedOlmEvent(
+                        content = megolmSession,
+                        sender = bob,
+                        senderKeys = keysOf(
+                            Ed25519Key(null, bobAccount.identityKeys.ed25519)
+                        ),
+                        recipient = alice,
+                        recipientKeys = keysOf(
+                            Ed25519Key(null, dehydratedDeviceAccount.identityKeys.ed25519)
+                        ),
+                    )
+                )
+            )
+        }
+        val encryptedMegolmSession =
+            OlmEncryptedToDeviceEventContent(
+                ciphertext = mapOf(
+                    dehydratedDeviceAccount.identityKeys.curve25519 to CiphertextInfo(
+                        encryptedMessage.cipherText,
+                        INITIAL_PRE_KEY
+                    )
+                ),
+                senderKey = KeyValue.Curve25519KeyValue(bobAccount.identityKeys.curve25519)
+            )
+
+        keyStore.updateDeviceKeys(bob) {
+            mapOf(
+                "BOB_DEVICE" to StoredDeviceKeys(
+                    SignedDeviceKeys(
+                        DeviceKeys(
+                            bob, "BOB_DEVICE", setOf(),
+                            keysOf(
+                                Ed25519Key("BOB_DEVICE", bobAccount.identityKeys.ed25519),
+                                Curve25519Key("BOB_DEVICE", bobAccount.identityKeys.curve25519),
+                            )
+                        ), mapOf()
+                    ),
+                    Valid(false)
+                )
+            )
+        }
+
+        apiConfig.endpoints {
+            matrixJsonEndpoint(GetDehydratedDevice()) {
+                getDehydratedDeviceCalled = true
+                GetDehydratedDevice.Response(
+                    deviceId = deviceId,
+                    deviceData = deviceData
+                )
+            }
+            matrixJsonEndpoint(GetDehydratedDeviceEvents(deviceId)) {
+                getDehydratedDeviceEventsCalled = true
+                if (it.nextBatch == null)
+                    GetDehydratedDeviceEvents.Response(
+                        events = listOf(ClientEvent.ToDeviceEvent(encryptedMegolmSession, bob)),
+                        nextBatch = "nextBatch1"
+                    )
+                else
+                    GetDehydratedDeviceEvents.Response(
+                        events = listOf(),
+                        nextBatch = "nextBatch2"
+                    )
+
+            }
+        }
+
+        dehydratedDeviceAccount.markKeysAsPublished()
+
+        cut.tryRehydrateDevice(SecureRandom.nextBytes(32))
+
+        getDehydratedDeviceCalled shouldBe true
+        getDehydratedDeviceEventsCalled shouldBe false
+        olmStore.getInboundMegolmSession(
+            megolmSession.sessionId,
+            megolmSession.roomId
+        ) shouldBe null
     }
 }